requirements.txt
docopt-ng~=0.9.0
PyMuPDF~=1.23.6
PyMuPDFb~=1.23.6
Pillow~=10.0.1